InventoryManagementSystem-main/requirements.txt
sqlite3
pillow==9.2.0